Split studio stories into per-file modules; colocate CSS; drop light mode
- Break each story out of the monolithic stories.tsx into its own file
  under src/studio/stories/ (registry kept in stories/index.tsx).
- Colocate styles with the component they style: each chrome component
  (Studio/StudioSidebar/StudioStage) and each story that needs styling
  (MenuStory/IconGalleryStory) owns a sibling .css.ts. Shared tokens stay
  in src/styles/settings. Removes the monolithic studio.css.ts.
- Remove the dark/light background toggle: the app has no light mode, so it
  was speculative and misrepresented the app. Studio now uses the app's
  actual dark background only.
