+++ src/studio/StudioStage.tsx
@@ -1,25 +1,18 @@
 import type { Story } from "./stories";
-import * as styles from "./studio.css";
+import * as styles from "./StudioStage.css";
 
 interface Props {
   story: Story;
-  light: boolean;
-  onToggleLight: () => void;
 }
 
-export const StudioStage = ({ story, light, onToggleLight }: Props) => {
+export const StudioStage = ({ story }: Props) => {
   const Component = story.Component;
   return (
     <main className={styles.stage}>
       <div className={styles.stageControls}>
         <span>{story.title}</span>
-        <button className={styles.controlButton} onClick={onToggleLight}>
-          {light ? "Dark bg" : "Light bg"}
-        </button>
       </div>
-      <div
-        className={`${styles.stageCanvas} ${light ? styles.stageCanvasLight : ""}`}
-      >
+      <div className={styles.stageCanvas}>
         <Component />
       </div>
     </main>
