+++ src/studio/stories/LoadingStory.tsx
@@ -1,0 +1,3 @@
+import { LoadingIcon } from "../../components/custom-song-icons/LoadingIcon";
+
+export const LoadingStory = () => <LoadingIcon />;
