+++ src/studio/stories/MenuStory.css.ts
@@ -1,0 +1,7 @@
+import { style } from "@vanilla-extract/css";
+
+export const menuStage = style({
+  position: "relative",
+  width: "100%",
+  height: "45rem",
+});
